refactor: move Transactions component to src/components
