+++ src/components/Transactions/TransactionDetail.scss
